Fetch client code based on info in a client-info file.
This makes it possible to bootstrap with a specific version of the
client code. The idea is that if future versions of the client become
incompatible with your project development environment, you can still
use a version that works.

Remove special-casing of asdf.lisp from the bootstrap process; it will
now be provided in quicklisp.tar.